api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
boto3
